Add support for x-ms-external extension
This allows types to be defined external to codegen.
