+++ test/maps/azalias/zz_generated_models_serde.go
@@ -191,37 +191,6 @@ func (g *GeoJSONObject) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type GeoJSONObjectNamedCollection.
-func (g GeoJSONObjectNamedCollection) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "collectionName", g.CollectionName)
-	populate(objectMap, "objects", g.Objects)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type GeoJSONObjectNamedCollection.
-func (g *GeoJSONObjectNamedCollection) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", g, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "collectionName":
-			err = unpopulate(val, "CollectionName", &g.CollectionName)
-			delete(rawMsg, key)
-		case "objects":
-			g.Objects, err = unmarshalGeoJSONObjectClassificationMap(val)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", g, err)
-		}
-	}
-	return nil
-}
-
 // MarshalJSON implements the json.Marshaller interface for type ListItem.
 func (l ListItem) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
